+++ mono/mind_maps/static/mind_maps/mind_map_detail.js
@@ -1,7 +1,6 @@
 var toolbar
 
 function init() {
-    initializePanel()
     loadMindMap()
     startSyncer()
     initToolbar()
@@ -12,9 +11,9 @@ function loadMindMap() {
         on: 'now',
         method: 'GET',
         url: `/mm/api/mind_maps/${MIND_MAP_ID}/`,
-        onSuccess: function (response) {
+        onSuccess(r) {
             nodes = []
-            response.nodes.forEach(n => {
+            r.nodes.forEach(n => {
                 let newNode = Node.getOrCreate(n.id)
                 newNode.name = n.name
                 newNode.position = n.position
@@ -30,9 +29,32 @@ function loadMindMap() {
                     newNode.parent = Node.getOrCreate(n.parent)
                 }
             })
-            renderNodes()
+            reposition(nodes)
+            centralize()
         }
     })
+}
+
+function reposition(nodes) {
+    const MARGIN = 20
+    let maxX = Math.max(...nodes.map(n => n.position[0]))
+    let minX = Math.min(...nodes.map(n => n.position[0]))
+    let width = maxX - minX
+    let maxY = Math.max(...nodes.map(n => n.position[1]))
+    let minY = Math.min(...nodes.map(n => n.position[1]))
+    let height = maxY - minY
+    let expandX = (width < window.innerWidth / scale) ? (window.innerWidth / scale - width) / 2 : 0
+    let expandY = (height < window.innerHeight / scale) ? (window.innerHeight / scale - height) / 2 : 0
+    let dX = MARGIN - minX + expandX
+    let dY = MARGIN - minY + expandY
+    nodes.forEach(n => {
+        n.position = [n.position[0] + dX, n.position[1] + dY]
+    })
+    initializePanel(
+        Math.max(window.innerWidth / scale, width) + 2 * MARGIN,
+        Math.max(window.innerHeight / scale, height) + 2 * MARGIN,
+    )
+    renderNodes()
 }
 
 function renderNodes() {
@@ -41,7 +63,6 @@ function renderNodes() {
         n.draw()
         n.drawConnectors()
     })
-    centralize()
     storeState()
 }